fixed searchSeries serialization
