+++ app/src/main/java/ninja/esgi/tvdbandroidapp/model/response/SearchSeriesResponse.java
@@ -6,99 +6,15 @@ import com.google.gson.annotations.SerializedName;
 import java.util.List;
 
 public class SearchSeriesResponse {
-    @SerializedName("aliases")
+    @SerializedName("data")
     @Expose
-    private List<String> aliases;
+    private List<SearchSeriesDataResponse> data;
 
-    @SerializedName("banner")
-    @Expose
-    private String banner;
-
-    @SerializedName("firstAired")
-    @Expose
-    private String firstAired;
-
-    @SerializedName("id")
-    @Expose
-    private Long id;
-
-    @SerializedName("network")
-    @Expose
-    private Long network;
-
-    @SerializedName("overview")
-    @Expose
-    private Long overview;
-
-    @SerializedName("seriesName")
-    @Expose
-    private Long seriesName;
-
-    @SerializedName("status")
-    @Expose
-    private Long status;
-
-    public List<String> getAliases() {
-        return aliases;
+    public List<SearchSeriesDataResponse> getData() {
+        return data;
     }
 
-    public void setAliases(List<String> aliases) {
-        this.aliases = aliases;
-    }
-
-    public String getBanner() {
-        return banner;
-    }
-
-    public void setBanner(String banner) {
-        this.banner = banner;
-    }
-
-    public String getFirstAired() {
-        return firstAired;
-    }
-
-    public void setFirstAired(String firstAired) {
-        this.firstAired = firstAired;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getNetwork() {
-        return network;
-    }
-
-    public void setNetwork(Long network) {
-        this.network = network;
-    }
-
-    public Long getOverview() {
-        return overview;
-    }
-
-    public void setOverview(Long overview) {
-        this.overview = overview;
-    }
-
-    public Long getSeriesName() {
-        return seriesName;
-    }
-
-    public void setSeriesName(Long seriesName) {
-        this.seriesName = seriesName;
-    }
-
-    public Long getStatus() {
-        return status;
-    }
-
-    public void setStatus(Long status) {
-        this.status = status;
+    public void setData(List<SearchSeriesDataResponse> data) {
+        this.data = data;
     }
 }
